main

+++ Helper/app/Http/Controllers/DonateController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Models\Donate;
 use Illuminate\Http\Request;
+use Session;
 
 class DonateController extends Controller
 {
@@ -14,8 +15,13 @@ class DonateController extends Controller
      */
     public function index()
     {
+        if (Session::has('loginId')) {
         $donates = Donate::all();
         return view('admin.donate.donateInfo')->with('donates', $donates);
+        }
+         else {
+            return view('admin.adminpages.login');
+        }
     }
 
     /**
@@ -36,23 +42,28 @@ class DonateController extends Controller
      */
     public function store(Request $request)
     {
-        $this->validate($request, [
-            'name' => 'required',
-            'email' => 'required',
-            'number' => 'required',
-            'address' => 'required',
-            'tool' => 'required',
-        ]);
+        if (Session::has('loginId')) {
+            $this->validate($request, [
+                'name' => 'required',
+                'email' => 'required',
+                'number' => 'required',
+                'address' => 'required',
+                'tool' => 'required',
+            ]);
 
-        $donate = new Donate();
-        $donate->name = $request->input('name');
-        $donate->email = $request->input('email');
-        $donate->number = $request->input('number');
-        $donate->address = $request->input('address');
-        $donate->tool = $request->input('tool');
-        $donate->save();
-        return redirect('/donate')->with('success', 'Donate Edited');
+            $donate = new Donate();
+            $donate->name = $request->input('name');
+            $donate->email = $request->input('email');
+            $donate->number = $request->input('number');
+            $donate->address = $request->input('address');
+            $donate->tool = $request->input('tool');
+            $donate->save();
+            return redirect('/donate')->with('success', 'Donate Edited');
 
+        }
+     else {
+            return view('admin.adminpages.login');
+        }
     }
 
     /**
@@ -74,9 +85,13 @@ class DonateController extends Controller
      */
     public function edit($id)
     {
-        $donate = Donate::find($id);
-        return view('admin.donate.donateEdit')->with('donate', $donate);
-
+        if (Session::has('loginId')) {
+            $donate = Donate::find($id);
+            return view('admin.donate.donateEdit')->with('donate', $donate);
+        }
+ else {
+            return view('admin.adminpages.login');
+        }
     }
 
     /**
@@ -88,16 +103,19 @@ class DonateController extends Controller
      */
     public function update(Request $request, $id)
     {
-        $donate = Donate::find($id);
-        // dd($donate);
-        $donate->name = $request->input('name');
-        $donate->email = $request->input('email');
-        $donate->number = $request->input('number');
-        $donate->address = $request->input('address');
-        $donate->tools = $request->input(['tools']);
-        $donate->save();
-        return redirect('/donate')->with('success', "Admin Edited");
-
+        if (Session::has('loginId')) {
+            $donate = Donate::find($id);
+            // dd($donate);
+            $donate->name = $request->input('name');
+            $donate->email = $request->input('email');
+            $donate->number = $request->input('number');
+            $donate->address = $request->input('address');
+            $donate->tools = $request->input(['tools']);
+            $donate->save();
+            return redirect('/donate')->with('success', "Admin Edited");
+        } else {
+            return view('admin.adminpages.login');
+        }
     }
 
     /**
@@ -108,9 +126,14 @@ class DonateController extends Controller
      */
     public function destroy($id)
     {
-        $donate = Donate::find($id);
-        $donate->delete();
-        return redirect('/donate')->with('success', "Admin Deleted");
+        if (Session::has('loginId')) {
+            $donate = Donate::find($id);
+            $donate->delete();
+            return redirect('/donate')->with('success', "Admin Deleted");
+        }
+         else {
+            return view('admin.adminpages.login');
+        }
 
     }
 }
